2nd upload
Follows line in some way (in the main function inside of the MainModule2 added _map() function)!!! But low FPS (stil not measured). Because of that it f!*#@ up in the turns!!! Speed is too high but you can't slow the motors otherwise the will stall. Trying to solve by moving it by steps of approximately .5s while allowing it to precess the imgs.